make LSB hiding algorithm working on Gupta-Garg scheme(where I hide k most significant bit of secret pixel in k least significant bits of carrier's pixel); some fixes in StegDetector; stegDetector now only calculate mean and deviation of Fx and Fy;

+++ StegDetector/LSBDetection.hpp
@@ -29,9 +29,9 @@ enum class LsbMode
 
 
 
-template<typename FwdIt> size_t sum(FwdIt First, FwdIt Last)
+template<typename FwdIt> int64_t sum(FwdIt First, FwdIt Last)
 {
-	size_t result = 0;
+	int64_t result = 0;
 	while (First != Last)
 	{
 		result += *First;
